requirements.txt
streamlit
tensorflow-cpu
pandas
requests
numpy
pickle5
streamlit-option-menu
Pillow==11.1.0